backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-pymongo>=2.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
bcrypt>=4.1.0
pyjwt>=2.8.0
werkzeug>=3.0.0
requests>=2.31.0
